+++ entrypoints/offscreen/database.js
@@ -15,6 +15,7 @@ import { ChangeLogV6 } from './changeLog/changeLogV6';
 import { ChangeLogV7 } from './changeLog/changeLogV7';
 import { ChangeLogV8 } from './changeLog/changeLogV8';
 import { ChangeLogV9 } from './changeLog/changeLogV9';
+import { ChangeLogV11 } from './changeLog/changeLogV11';
 import { postErrorMessage, postSuccessMessage } from "./util";
 
 const JOB_DB_FILE_NAME = "job.sqlite3";
@@ -498,6 +499,7 @@ export const Database = {
       changelogList.push(new ChangeLogV8());
       changelogList.push(new ChangeLogV9());
       changelogList.push(new ChangeLogV10());
+      changelogList.push(new ChangeLogV11());
       initChangeLog(changelogList);
       sqlite3InitModule({
         print: debugLog,